Fixed a bug where monsters were not instanced
The monster objects on the prefabs were being changed instead of
instances of those objects. Had to remove those from the prefab and put
the monster component on the same object. Then I initialize a monster in
code (no more monster prefabs). As a result, I also had to fix the
GameController to be a static object instead of a gameObject that wasnt
destroyed