New Map Asset, new Player sprite, map scene wip
Using a new map asset that seems better. It has a built-in way to tie
real world lat/lon to unity space. Nothing spawns anymore on the map and
moving is different. Started tying to device so drag to move but it
makes the map freak out. Need to fix it for 2D, I think. Changed aspect
ratio of game to iPhone. Drastically changed the scale of th map scene
to try and make the map work. Added a bunch of prefabs of old map
objects.