Map still wip
Tried some things but the map is still wonky. It loads way sooner than
it should and it jumps the character really far when it switches because
it's not lining up right or something. Disabled loading new map tiles
for now. Lat/lon to unity position seems to be the real problem but I
still havent figured out how/why