Hacked map to line up with game space
Changed a couple weird, small things about the map to make it work. For
some reason the area that it thinks is world space is about 8 times
smaller than it should be. I dont know what is making it too small, but
just scaling it up lined things up more closely. It's not perfect
because I dont know exactly how far off it is. It just happens to be
about 8. Leaving it this way for now.