Refactrored a bunch of stuff and improved Monster
Moved a bunch of stuff around in the folder structure. Scripts are
separated by scenes. Prefabs too and I moevd them to a folder called
Resources. I also cleaned up unused or unecessary prefabs. Now they are
loaded in code instead of in the inspector. This lets me set up the
Monster object to load GameObjects for its Fight and Capture scenes.
Everything should be working same as before, otherwise.