More pinball updates for performance
I made the capture monster stay in one place and teleport around. This
is how I want it and not just for performance. Eventually they will have
idle animations and stuff, but for now I switched them to actually use
animation controllers so it will be easier to integrate the various
future animations.