Fight Scene Updates
Updated the fight scene to show all stats and load the player's team in
the team pane. Also refactored a bunch of Monster stuff to flesh it out
more. Still dont have a good way to handle Moves and stuff. Need to
start thinking about a datasource for all the monster, move, etc stuff.
Fight scene can now be run directly and it will populate with a couple
monsters for easier testing